refactor controller initializer
