+++ src/decorators/controller/controller.ts
@@ -1,16 +1,10 @@
-import normalizePath from "normalize-path";
 import { container as defaultContainer } from "../../container/container";
 import { keys } from "../../container/keys";
-import { getGlobal } from "../../container/utils/getGlobal";
 import { validateContainer } from "../../container/utils/validateContainer";
-import { setInjectables } from "../inject/setInjectables";
-import type { Class, ClassDecorator, HttpMethodMetadata } from "../types";
-import type { AnnotatedRouter } from "../../interfaces/annotatedRouter";
+import type { Class, ClassDecorator } from "../types";
 import { addClassToContainer } from "../utils/addClassToContainer";
-import { getMetadata } from "../utils/getMetadata";
-import { getMetadataProperty } from "../utils/getMetadataProperty";
 import { validateKind } from "../utils/validateKind";
-import { MetadataProperties } from "./metadataProperties";
+import { getInitializer } from "./getInitializer";
 
 /**
  * A class decorator that encapsulates API endpoints
@@ -85,43 +79,9 @@ export const Controller = <T extends Class<object>>(
       throw new Error("Controller path argument is an empty string");
     }
 
-    context.addInitializer(function () {
-      const controller = new this();
-
-      const metadata = getMetadata(annotationName, context);
-      setInjectables(container, controller, metadata);
-
-      const methods = <Array<HttpMethodMetadata>>(
-        getMetadataProperty(metadata, MetadataProperties.methods, [])
-      );
-
-      const controllerMethodMetadata = methods.map((methodMetadata) => ({
-        ...methodMetadata,
-        handler: methodMetadata.handler.bind(controller),
-      }));
-
-      let router: AnnotatedRouter = getGlobal(container, keys.router);
-      for (const {
-        path: methodPath,
-        handler,
-        httpMethod,
-      } of controllerMethodMetadata) {
-        const routerMethod = <Exclude<keyof AnnotatedRouter, "handle">>(
-          httpMethod.toLowerCase()
-        );
-
-        if (typeof router[routerMethod] !== "function") {
-          throw new Error(
-            `Router is improperly configured. It should include ${routerMethod} method`,
-          );
-        }
-
-        router = router[routerMethod](
-          normalizePath("/" + [controllerPath, methodPath].join("/"), true),
-          handler,
-        );
-      }
-    });
+    context.addInitializer(
+      getInitializer(annotationName, context, container, controllerPath),
+    );
 
     addClassToContainer(container, keys.controllerClasses, constructor);
   }) as ClassDecorator<T>;
